UPSTREAM: vsock: handle buffer_size sockopts in the core
virtio_transport and vmci_transport handle the buffer_size
sockopts in a very similar way.

In order to support multiple transports, this patch moves this
handling in the core to allow the user to change the options
also if the socket is not yet assigned to any transport.

This patch also adds the '.notify_buffer_size' callback in the
'struct virtio_transport' in order to inform the transport,
when the buffer_size is changed by the user. It is also useful
to limit the 'buffer_size' requested (e.g. virtio transports).

Acked-by: Dexuan Cui <decui@microsoft.com>
Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Reviewed-by: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit b9f2b0ffde0c9b666b2b1672eb468b8f805a9b97)

BUG=b:153004946
TEST=nested vsock works

Change-Id: I8869595fe467f31eddb89e8566fbbc66879b5cab
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2406623
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
[5.4-arcvm: picked from 5.4 for initial merge]