UPSTREAM: vsock: prevent transport modules unloading
This patch adds 'module' member in the 'struct vsock_transport'
in order to get/put the transport module. This prevents the
module unloading while sockets are assigned to it.

We increase the module refcnt when a socket is assigned to a
transport, and we decrease the module refcnt when the socket
is destructed.

Reviewed-by: Stefan Hajnoczi <stefanha@redhat.com>
Reviewed-by: Jorgen Hansen <jhansen@vmware.com>
Signed-off-by: Stefano Garzarella <sgarzare@redhat.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 6a2c0962105ae8ceba182c4f616e0e41d7755591)

BUG=b:153004946
TEST=nested vsock works

Change-Id: I37167c07f5b8fdfca7c561415cb172c12e24108c
Signed-off-by: Allen Webb <allenwebb@google.com>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2406628
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Sean Paul <seanpaul@chromium.org>
[5.4-arcvm: picked from 5.4 for initial merge]