BACKPORT: dax: Pass dax_dev instead of bdev to dax_writeback_mapping_range()
As of now dax_writeback_mapping_range() takes "struct block_device" as a
parameter and dax_dev is searched from bdev name. This also involves taking
a fresh reference on dax_dev and putting that reference at the end of
function.

We are developing a new filesystem virtio-fs and using dax to access host
page cache directly. But there is no block device. IOW, we want to make
use of dax but want to get rid of this assumption that there is always
a block device associated with dax_dev.

So pass in "struct dax_device" as parameter instead of bdev.

ext2/ext4/xfs are current users and they already have a reference on
dax_device. So there is no need to take reference and drop reference to
dax_device on each call of this function.

Suggested-by: Christoph Hellwig <hch@infradead.org>
Reviewed-by: Christoph Hellwig <hch@lst.de>
Reviewed-by: Jan Kara <jack@suse.cz>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Link: https://lore.kernel.org/r/20200103183307.GB13350@redhat.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

(cherry picked from commit 3f666c56c6b8cc40a5e9002aac484b8f5b83c402)

Conflict:
  fs/xfs/xfs_aops.c

This commit has to be cherry picked because "virtiofs: add DAX mmap
support" depends on the change of dax_writeback_mapping_range().

BUG=b:153701115
TEST=build

[5.4-arcvm: picked from 5.4]
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Change-Id: I8bfbf079421ffeeb0d4a7fcb126e06b8114d9997
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2536228
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2596194
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>