UPSTREAM: dax, pmem: Add a dax operation zero_page_range
Add a dax operation zero_page_range, to zero a page. This will also clear any
known poison in the page being zeroed.

As of now, zeroing of one page is allowed in a single call. There
are no callers which are trying to zero more than a page in a single call.
Once we grow the callers which zero more than a page in single call, we
can add that support. Primary reason for not doing that yet is that this
will add little complexity in dm implementation where a range might be
spanning multiple underlying targets and one will have to split the range
into multiple sub ranges and call zero_page_range() on individual targets.

Suggested-by: Christoph Hellwig <hch@infradead.org>
Signed-off-by: Vivek Goyal <vgoyal@redhat.com>
Reviewed-by: Pankaj Gupta <pankaj.gupta.linux@gmail.com>
Link: https://lore.kernel.org/r/20200228163456.1587-3-vgoyal@redhat.com
Signed-off-by: Dan Williams <dan.j.williams@intel.com>

(cherry picked from commit f605a263e0690177ecc180417eacf2b5507dd177)

BUG=b:153701115
TEST=build

[5.4-arcvm: picked from 5.4]
Signed-off-by: Keiichi Watanabe <keiichiw@chromium.org>
Change-Id: I73f9b78c5c6a6f9d0d1908be02791476af91675d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2550867
Reviewed-by: Sean Paul <seanpaul@chromium.org>
Reviewed-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2596189