CHROMIUM: uapi/fuse.h: Add message definitions for CHROMEOS_TMPFILE
This is a chrome os extension for fuse implemented in a way that will
hopefully not conflict with rebases.  The upstream maintainer was not
happy with this approach because it uses 2 round-trips instead of one.
However, his proposed mechanism requires changes to the VFS layer but
the VFS maintainer was non-committal.

At this point, I think the patch is non-intrusive enough that we can
just maintain it ourselves and if upstream ever gains support for
O_TMPFILE we can drop it at that time.

BUG=b:160932094
TEST=Successfully open a file with  O_TMPFILE on virtio-fs

[5.4-arcvm: picked from 5.4]
Change-Id: Ieaa10fbe89a143c3f39a4fd73c60378e7f57bda2
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2520739
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
(cherry picked from commit 70be01c0cc21fb267f54195d3d9572a2143c6c30)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2581186
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>