CHROMIUM: fuse: Implement CHROMEOS_TMPFILE
This is a chrome os extension for fuse implemented in a way that will
hopefully not conflict with rebases.  The upstream maintainer was not
happy with this approach because it uses 2 round-trips instead of one.
However, his proposed mechanism requires changes to the VFS layer but
the VFS maintainer was non-committal.

At this point, I think the patch is non-intrusive enough that we can
just maintain it ourselves and if upstream ever gains support for
O_TMPFILE we can drop it at that time.

BUG=b:160932094
TEST=Successfully open a file with  O_TMPFILE on virtio-fs

[5.4-arcvm: picked from 5.4]
Change-Id: I600abd2d7604f8e167b6be65f5194e18bbf1b554
Signed-off-by: Chirantan Ekbote <chirantan@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2520740
Reviewed-by: Suleiman Souhlal <suleiman@chromium.org>
(cherry picked from commit b7eaf1252a818d0b557786b76616fe112b2d7547)
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/third_party/kernel/+/2581136
Reviewed-by: Hikaru Nishida <hikalium@chromium.org>
Commit-Queue: Kazuhiro Inaba <kinaba@chromium.org>